backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
kubernetes==29.0.0
pyyaml==6.0.1
python-dotenv==1.0.0
jinja2==3.1.2